# perf(object-hash): avoid unnecessary `.toString()` on known values